add minitest whitelist feature if running with rails 6.0
The minitest whitelist feature can be used to allow test to fail that
are on the test/whitelist.yml. The whitelist is only active when
minitests is running with rails version 6.0 or higher.

The minitest testreportes work as expected and report the tests as
failing but the minitest exit code is manipulated and 0 if all tests are
passing or the failing tests are on the whitelist.

The exit code is 1 if there are additional failures.

The junit reporter creates the normal xml output and the ci pipeline
might fail if it uses the test reports to determine if the testsuit run
was successfull. Please use only the exit code.